+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/StockLocationServiceSupplychain.java
@@ -18,9 +18,10 @@
 package com.axelor.apps.supplychain.service;
 
 import com.axelor.apps.stock.service.StockLocationService;
+import com.axelor.exception.AxelorException;
 import java.math.BigDecimal;
 
 public interface StockLocationServiceSupplychain extends StockLocationService {
 
-  public BigDecimal getReservedQty(Long productId, Long locationId);
+  public BigDecimal getReservedQty(Long productId, Long locationId) throws AxelorException;
 }